requirement.txt
django==4.0.1
django_crispy_forms
django-filter
mysqlclient
Pillow